feat: commit isConditionalExpression generateCode
