+++ eslint.config.js
@@ -1,13 +1,17 @@
 import { icebreaker } from '@icebreakers/eslint-config'
 
-export default icebreaker({
-  // mdx: true,
-}, {
-  ignores: [
-    '**/fixtures/**',
-    'apps',
-    'demo',
-    'demo-linked',
-    'how-to-build-components-by-tailwindcss',
-  ],
-})
+export default icebreaker(
+  {
+    markdown: false,
+    // mdx: true,
+  },
+  {
+    ignores: [
+      '**/fixtures/**',
+      'apps',
+      'demo',
+      'demo-linked',
+      'how-to-build-components-by-tailwindcss',
+    ],
+  },
+)
