+++ packages/weapp-tailwindcss/test/js.test.ts
@@ -9,12 +9,11 @@ import { TailwindcssPatcher } from 'tailwindcss-patch'
 
 // swc 在解析中文的时候，会导致 span 增加，从而无法精确定位，不知道是不是bug
 import { createGetCase, createPutCase, jsCasePath, tsCasePath } from './util'
-import { SimpleMappingChars2String } from '@/escape'
+import { SimpleMappingChars2String, decodeUnicode } from '@/escape'
 import { createJsHandler } from '@/js/index'
 import { getCss } from '#test/helpers/getTwCss'
 import { getOptions } from '@/options'
 import { defaultOptions } from '@/defaults'
-import { decodeUnicode } from '@/js/handlers'
 // import { parse } from '@/babel'
 const getCase = createGetCase(jsCasePath)
 const getTsCase = createGetCase(tsCasePath)
